Update x0 value to same size as x1, x2, x3, etc.
The code for the 2nd, 3rd, and 4th iterations is identical. By making
the x0 value the same size as x1, x2, and x3 the code for all of the
iterations is effectively identical. This means tha I can more
efficiently use either a `for' loop or a state machine to implement the
circuit.

And the code does need to use a state machine. In its current
configuration the circuit is pure combinatorial logic which is
inadequate for a multi-cycle instruction.