Replace "if" statement with a "for" loop.
The conditional logic to support all of the IEEE 754 binary formats was
too complicated. Because the width of x0 is now the same as 1, x2, and
x3 the "if" statement can be replaced with a "for" loop making the code
much simpler, and more compact.

Also, using the "for" loop takes us one step closer converting this
combinatorial logic into a statement machine.